+++ src/paging.rs
@@ -275,7 +275,8 @@ impl<T: Translation> RootTable<T> {
     }
 
     /// Recursively maps a range into the pagetable hierarchy starting at the root level, mapping
-    /// the pages to the corresponding physical address range starting at `pa`.
+    /// the pages to the corresponding physical address range starting at `pa`. Block and page
+    /// entries will be written to, but will only be mapped if `flags` contains `Attributes::VALID`.
     ///
     /// Returns an error if the virtual address range is out of the range covered by the pagetable
     pub fn map_range(
@@ -469,7 +470,8 @@ impl<T: Translation> PageTableWithLevel<T> {
     }
 
     /// Maps the the given virtual address range in this pagetable to the corresponding physical
-    /// address range starting at the given `pa`, recursing into any subtables as necessary.
+    /// address range starting at the given `pa`, recursing into any subtables as necessary. To map
+    /// block and page entries, `Attributes::VALID` must be set in `flags`.
     ///
     /// Assumes that the entire range is within the range covered by this pagetable.
     ///
@@ -506,19 +508,22 @@ impl<T: Translation> PageTableWithLevel<T> {
                 } else {
                     let old = *entry;
                     let (mut subtable, subtable_pa) = Self::new(translation, level + 1);
-                    if let (Some(old_flags), Some(old_pa)) = (old.flags(), old.output_address()) {
-                        // Old was a valid block entry, so we need to split it.
-                        // Recreate the entire block in the newly added table.
-                        let a = align_down(chunk.0.start.0, granularity);
-                        let b = align_up(chunk.0.end.0, granularity);
-                        subtable.map_range(
-                            translation,
-                            &MemoryRegion::new(a, b),
-                            old_pa,
-                            old_flags,
-                        );
+                    if let Some(old_flags) = old.flags() {
+                        if !old_flags.contains(Attributes::TABLE_OR_PAGE) {
+                            let old_pa = old.output_address();
+                            // `old` was a block entry, so we need to split it.
+                            // Recreate the entire block in the newly added table.
+                            let a = align_down(chunk.0.start.0, granularity);
+                            let b = align_up(chunk.0.end.0, granularity);
+                            subtable.map_range(
+                                translation,
+                                &MemoryRegion::new(a, b),
+                                old_pa,
+                                old_flags,
+                            );
+                        }
                     }
-                    entry.set(subtable_pa, Attributes::TABLE_OR_PAGE);
+                    entry.set(subtable_pa, Attributes::TABLE_OR_PAGE | Attributes::VALID);
                     subtable
                 };
                 subtable.map_range(translation, &chunk, pa, flags);
@@ -533,6 +538,7 @@ impl<T: Translation> PageTableWithLevel<T> {
         translation: &T,
         indentation: usize,
     ) -> Result<(), fmt::Error> {
+        const WIDTH: usize = 3;
         // Safe because we know that the pointer is aligned, initialised and dereferencable, and the
         // PageTable won't be mutated while we are using it.
         let table = unsafe { self.table.as_ref() };
@@ -545,12 +551,16 @@ impl<T: Translation> PageTableWithLevel<T> {
                     i += 1;
                 }
                 if i - 1 == first_zero {
-                    writeln!(f, "{:indentation$}{}: 0", "", first_zero)?;
+                    writeln!(f, "{:indentation$}{: <WIDTH$}: 0", "", first_zero)?;
                 } else {
-                    writeln!(f, "{:indentation$}{}-{}: 0", "", first_zero, i - 1)?;
+                    writeln!(f, "{:indentation$}{: <WIDTH$}-{}: 0", "", first_zero, i - 1)?;
                 }
             } else {
-                writeln!(f, "{:indentation$}{}: {:?}", "", i, table.entries[i])?;
+                writeln!(
+                    f,
+                    "{:indentation$}{: <WIDTH$}: {:?}",
+                    "", i, table.entries[i],
+                )?;
                 if let Some(subtable) = table.entries[i].subtable(translation, self.level) {
                     subtable.fmt_indented(f, translation, indentation + 2)?;
                 }
@@ -653,12 +663,8 @@ pub struct Descriptor(usize);
 impl Descriptor {
     const PHYSICAL_ADDRESS_BITMASK: usize = !(PAGE_SIZE - 1) & !(0xffff << 48);
 
-    fn output_address(self) -> Option<PhysicalAddress> {
-        if self.is_valid() {
-            Some(PhysicalAddress(self.0 & Self::PHYSICAL_ADDRESS_BITMASK))
-        } else {
-            None
-        }
+    fn output_address(self) -> PhysicalAddress {
+        PhysicalAddress(self.0 & Self::PHYSICAL_ADDRESS_BITMASK)
     }
 
     /// Returns the flags of this page table entry, or `None` if its state does not
@@ -687,7 +693,7 @@ impl Descriptor {
     }
 
     fn set(&mut self, pa: PhysicalAddress, flags: Attributes) {
-        self.0 = (pa.0 & Self::PHYSICAL_ADDRESS_BITMASK) | (flags | Attributes::VALID).bits();
+        self.0 = (pa.0 & Self::PHYSICAL_ADDRESS_BITMASK) | flags.bits();
     }
 
     fn subtable<T: Translation>(
@@ -696,10 +702,9 @@ impl Descriptor {
         level: usize,
     ) -> Option<PageTableWithLevel<T>> {
         if level < LEAF_LEVEL && self.is_table_or_page() {
-            if let Some(output_address) = self.output_address() {
-                let table = translation.physical_to_virtual(output_address);
-                return Some(PageTableWithLevel::from_pointer(table, level + 1));
-            }
+            let output_address = self.output_address();
+            let table = translation.physical_to_virtual(output_address);
+            return Some(PageTableWithLevel::from_pointer(table, level + 1));
         }
         None
     }
@@ -708,8 +713,10 @@ impl Descriptor {
 impl Debug for Descriptor {
     fn fmt(&self, f: &mut Formatter) -> Result<(), fmt::Error> {
         write!(f, "{:#016x}", self.0)?;
-        if let (Some(flags), Some(address)) = (self.flags(), self.output_address()) {
-            write!(f, " ({}, {:?})", address, flags)?;
+        if self.is_valid() {
+            if let Some(flags) = self.flags() {
+                write!(f, " ({}, {:?})", self.output_address(), flags)?;
+            }
         }
         Ok(())
     }
@@ -831,18 +838,19 @@ mod tests {
 
     #[test]
     fn set_descriptor() {
+        const PHYSICAL_ADDRESS: usize = 0x12340000;
         let mut desc = Descriptor(0usize);
         assert!(!desc.is_valid());
         desc.set(
-            PhysicalAddress(0x12340000),
-            Attributes::TABLE_OR_PAGE | Attributes::USER | Attributes::SWFLAG_1,
+            PhysicalAddress(PHYSICAL_ADDRESS),
+            Attributes::TABLE_OR_PAGE | Attributes::USER | Attributes::SWFLAG_1 | Attributes::VALID,
         );
         assert!(desc.is_valid());
         assert_eq!(
             desc.flags().unwrap(),
             Attributes::TABLE_OR_PAGE | Attributes::USER | Attributes::SWFLAG_1 | Attributes::VALID
         );
-        assert_eq!(desc.output_address().unwrap(), PhysicalAddress(0x12340000));
+        assert_eq!(desc.output_address(), PhysicalAddress(PHYSICAL_ADDRESS));
     }
 
     #[test]